OpenAI_SDK_test/pyproject.toml
[project]
name = "openai-sdk-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "ipykernel>=6.31.0",
    "openai-agents>=0.6.8",
    "torch",
    "torchvision",
    "transformers>=4.57.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.31.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
